pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "num-agents"
version = "0.1.0"
description = "Nüm Agents: A dynamic agent orchestration framework"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "NUMTEMA", email = "info@numtema.com" }
]
keywords = ["agents", "ai", "orchestration", "framework", "num"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "pyyaml>=6.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.0.1",
    "mypy>=1.0.0",
]

[project.scripts]
num-agents = "num_agents.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["num_agents"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "B"]
ignore = []

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
